requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
aiosqlite==0.21.0
beautifulsoup4==4.12.2
gspread==5.12.4
httpx==0.27.0
lxml[html_clean]==4.9.3
oauth2client==4.1.3
openai==1.51.0
Pillow==10.1.0
python-dotenv==1.0.0
reportlab==4.0.8
sqlalchemy==2.0.30
aiofiles==23.2.1